extract resolutions and specializations
