+++ vine/src/components/specializer.rs
@@ -1,15 +1,14 @@
-use std::collections::{hash_map::Entry, HashMap};
+use std::collections::hash_map::Entry;
 
-use vine_util::{idx::IdxVec, new_idx};
+use vine_util::idx::IdxVec;
 
-use crate::{
-  components::resolver::{Fragment, FragmentId, Resolutions},
-  structures::{
-    chart::{Chart, ConstId, FnId, ImplId},
-    checkpoint::Checkpoint,
-    diag::ErrorGuaranteed,
-    tir::{ConstRelId, FnRelId, TirImpl},
-  },
+use crate::structures::{
+  chart::{Chart, ConstId, FnId},
+  checkpoint::Checkpoint,
+  diag::ErrorGuaranteed,
+  resolutions::{Fragment, FragmentId, Resolutions},
+  specializations::{ImplTree, Spec, SpecId, SpecRels, Specializations},
+  tir::TirImpl,
 };
 
 #[derive(Debug)]
@@ -18,12 +17,6 @@ pub struct Specializer<'core, 'a> {
   pub resolutions: &'a Resolutions,
   pub specs: &'a mut Specializations,
   pub fragments: &'a IdxVec<FragmentId, Fragment<'core>>,
-}
-
-#[derive(Debug, Default)]
-pub struct Specializations {
-  pub lookup: IdxVec<FragmentId, HashMap<Vec<ImplTree>, SpecId>>,
-  pub specs: IdxVec<SpecId, Option<Spec>>,
 }
 
 impl<'core, 'a> Specializer<'core, 'a> {
@@ -100,25 +93,6 @@ impl<'core, 'a> Specializer<'core, 'a> {
     }
   }
 }
-
-new_idx!(pub SpecId);
-
-#[derive(Debug)]
-pub struct Spec {
-  pub fragment: FragmentId,
-  pub index: usize,
-  pub singular: bool,
-  pub rels: SpecRels,
-}
-
-#[derive(Debug)]
-pub struct SpecRels {
-  pub fns: IdxVec<FnRelId, Result<SpecId, ErrorGuaranteed>>,
-  pub consts: IdxVec<ConstRelId, Result<SpecId, ErrorGuaranteed>>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct ImplTree(ImplId, Vec<ImplTree>);
 
 fn instantiate(impl_: &TirImpl, params: &[ImplTree]) -> Result<ImplTree, ErrorGuaranteed> {
   match &impl_ {
