+++ vine/src/structures.rs
@@ -3,7 +3,9 @@ pub mod chart;
 pub mod checkpoint;
 pub mod core;
 pub mod diag;
+pub mod resolutions;
 pub mod signatures;
+pub mod specializations;
 pub mod tir;
 pub mod types;
 pub mod vir;
