+++ vine/src/tools/repl.rs
@@ -21,13 +21,14 @@ use crate::{
     analyzer::usage::Usage,
     charter::{Charter, ExtractItems},
     parser::VineParser,
-    resolver::{FragmentId, Resolver},
+    resolver::Resolver,
   },
   structures::{
     ast::{visit::VisitMut, Block, Ident, Span, Stmt},
     chart::{DefId, VariantId},
     core::Core,
     diag::Diag,
+    resolutions::FragmentId,
     tir::Local,
     types::{Type, TypeKind, Types},
     vir::{StageId, Vir},
